requirements.txt
psycopg2==2.8.3
sqlalchemy==1.3.5
requests==2.22.0
python-dotenv==0.10.3
fire==0.1.3
